refactor(repo): extract Namespace into separate module
